feat: 저장된 링크가 없는 경우 디자인 구현
