+++ src/services/api.jsx
@@ -5,5 +5,10 @@ export function getProfiles() {
 }
 
 export function getFolders() {
-  return fetch(`${API_URL}/sample/folder`).then((res) => res.json());
+  return fetch(`${API_URL}/users/1/folders`).then((res) => res.json());
+}
+
+export async function getLinks(folderId = "id") {
+  const query = `folderId=${folderId}`;
+  return fetch(`${API_URL}/users/1/links?${query}`).then((res) => res.json());
 }
